Move TransportStatusError in errors
